backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.11.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
psutil==5.9.8
supabase
openai>=1.12.0
tenacity==8.2.3
python-multipart==0.0.9
pydantic-settings==2.1.0
langgraph>=0.0.19
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
yfinance
pinecone>=3.0.0
alpha_vantage
sec-edgar-downloader
praw
asyncpg>=0.29.0
greenlet>=2.0
pyjwt
sse-starlette>=1.4.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
numpy<2.0.0
asyncpraw>=7.7.0
sec-parser>=0.2.6